+++ apps/frontend/src/components/overview/details.module.scss
