+++ apps/frontend/src/components/overview/details.tsx
